Add buy/delete/create ticket

+++ aviasales2/src/main/java/com/example/aviasales2/repository/WalletRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 @Repository
 public interface WalletRepository extends CrudRepository<Wallet, Long> {
 
-    Wallet findById(Integer id);
+    Wallet findById(long id);
 }
